ML/requirements.txt
pandas
numpy
langchain
newspaper3k
sentence_transformers
faiss-cpu
google.generativeai
fastapi
uvicorn
